02/requirements.txt
astroid==3.0.0
coverage==7.3.2
dill==0.3.7
factory-boy==3.3.0
Faker==19.10.0
isort==5.12.0
mccabe==0.7.0
platformdirs==3.11.0
pylint==3.0.1
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.8.0
